+++ src/reagent_dev_tools/context.cljs
@@ -1,0 +1,7 @@
+(ns reagent-dev-tools.context
+  (:require ["react" :as react]))
+
+(defonce panel-context (react/createContext nil))
+
+(def panel-context-provider (.-Provider panel-context))
+;; (def panel-context-consumer (.-Consumer panel-context))